+++ backend/src/media/media.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
 import { MinioService } from './minio.service';
 import { MediaService } from './media.service';
+import { ClamAvService } from './clamav.service';
 import { MediaController } from './media.controller';
 
 @Module({
   controllers: [MediaController],
-  providers: [MinioService, MediaService],
+  providers: [MinioService, MediaService, ClamAvService],
   exports: [MediaService, MinioService],
 })
 export class MediaModule {}